Removed the practice file
